requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
langchain-community
google-api-python-client
python-dotenv
groq
openai
faiss-cpu
pandas
plotly
matplotlib
google-generativeai
nest_asyncio